Fix mismatch of fprintf format specifier and argument type
This fixes a warning about a mismatch between a fprintf format string
and an argument type on Darwin-64-bit builds:

%lu specifies type 'unsigned long' but the argument has type
'__darwin_suseconds_t' (aka 'int')

Acked-by: Arne Schwabe <arne@rfc2549.org>
Message-Id: <CAEsd45SoC+h5tcDraZcQN_MUYzqWYOVvNpkwMo190DLpB32StQ@mail.gmail.com>
URL: http://article.gmane.org/gmane.network.openvpn.devel/9446

+++ src/openvpn/error.c
@@ -319,7 +319,7 @@ void x_msg_va (const unsigned int flags, const char *format, va_list arglist)
 
 	      fprintf (fp, "%lu.%06lu %x %s%s%s%s",
 		       tv.tv_sec,
-		       tv.tv_usec,
+		       (unsigned long)tv.tv_usec,
 		       flags,
 		       prefix,
 		       prefix_sep,
